Two more NSUInteger fixes.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@7982 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Frameworks/OgreKit/Source/FindPanel/OgreFindPanel.m
@@ -65,7 +65,7 @@ static OgreFindPanel *sharedInstance;
 }
 
 - (id) initWithContentRect: (NSRect) frame
-		 styleMask: (unsigned int) mask
+		 styleMask: (NSUInteger) mask
 		   backing: (NSBackingStoreType) type
 		     defer: (BOOL) defer
 {
